refactor: rename auth to login page
